+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [1.9.0] - 2023-09-27 (Multilanguage)
+### Added
+- Redirects app
+- language.js
+- Languages settings
+### Changed
+- URLs to i18n patterns
+- Flatpages URL to /pages/
+
 ## [1.8.9] - 2023-09-12
 ### Fixed
 - Fix fantasycopy bug